+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `Totalizer` to `EXPERIMENTAL_Table`.
+
 ## [9.105.1] - 2020-01-17
 
 ### Added